refactor(shortcuts): consolidate and simplify music control shortcuts
