+++ lib/components/Shortcuts/global_shortcut_manager.dart
@@ -1,5 +1,4 @@
-import 'package:finamp/components/Shortcuts/music_control_actions.dart';
-import 'package:finamp/components/Shortcuts/music_control_intents.dart';
+import 'package:finamp/components/Shortcuts/music_control_shortcuts.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 
@@ -12,15 +11,22 @@ class GlobalShortcutManager extends StatelessWidget {
   Widget build(BuildContext context) {
     return Shortcuts(
       shortcuts: {
+        // Music control shortcuts
         const SingleActivator(LogicalKeyboardKey.space): const TogglePlaybackIntent(),
         const SingleActivator(LogicalKeyboardKey.keyN, control: true): const SkipToNextIntent(),
         const SingleActivator(LogicalKeyboardKey.keyP, control: true): const SkipToPreviousIntent(),
-        const SingleActivator(LogicalKeyboardKey.arrowRight, control: true): const SeekForwardIntent(),
-        const SingleActivator(LogicalKeyboardKey.arrowLeft, control: true): const SeekBackwardIntent(),
-        const SingleActivator(LogicalKeyboardKey.arrowUp, control: true): const VolumeUpIntent(),
-        const SingleActivator(LogicalKeyboardKey.arrowDown, control: true): const VolumeDownIntent(),
+        const SingleActivator(LogicalKeyboardKey.arrowRight, control: true): const SeekForwardIntent(), // Seek +30s
+        const SingleActivator(LogicalKeyboardKey.arrowLeft, control: true): const SeekBackwardIntent(), // Seek -5s
+        const SingleActivator(LogicalKeyboardKey.arrowUp, control: true): const VolumeUpIntent(), // Vol +5%
+        const SingleActivator(LogicalKeyboardKey.arrowDown, control: true): const VolumeDownIntent(), // Vol -5%
       },
-      child: MusicControlActions(child: child),
+      child: Actions(
+        actions: {
+          ...getMusicControlActions(),
+          // Other actions can be added here
+        },
+        child: child,
+      ),
     );
   }
 }
